feat(cli): Simplified, cleaned scaffolding
Instead of scaffolding from strings, scaffolding is done from template files and through JSON
stringification. The Gruntfile has also been simplified and all ngbp-based config moved to a JSON
file for two reasons: it's easier to templatize complex information, and it will be easier to
manipulate with CLI tools, such as disabling modules or injecting tasks.