feat(cli): Added config-related manipulation flags
Config parameters can be manipulated at runtime using the generic --"prop" and "--set" or
"--replace" to interact with a key. "--write" will save the config changes to the ngbp.json file.
Speaking of which, ngbp-related config variables are no longer stored in Grunt, but in an
independent JSON file that will be injected into Grunt before tasks are executed; this is
beneficial both for easing migration to a potential post-Grunt release of ngbp as well as for
saving changes made during runtime.

There are also several flags to ease the manipulation of specific config variables. "--prevent" and
"--allow" prevent and allow specific injected tasks from running and "--inject" and
"--inject-watch" inject new tasks into hooks.

"--raw" returns config variables without processing them as templates and "--json" will process new
or altered config parameters as json before saving them, allowing the user to specify arrays or
objects via the command line